+++ Changes.md
@@ -1,7 +1,10 @@
 1.5.x.x (relative to 1.5.11.0)
 =======
 
+API
+---
 
+- PlugAlgo : Added support in `setValueFromData()` for setting StringPlug values from StringVectorData.
 
 1.5.11.0 (relative to 1.5.10.1)
 ========